requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
sqlalchemy
psycopg2-binary
